+++ python/clu/testing.py
@@ -63,7 +63,7 @@ class MockReply(dict):
 class MockReplyList(list):
     """Stores replies as `.MockReply` objects."""
 
-    LEGACY_REPLY_PATTERN = re.compile(r'([0-9]+)\s+([0-9]+)\s+((?:[a-z]|\:))\s+(.*)')
+    LEGACY_REPLY_PATTERN = re.compile(r'([0-9]+)\s+([0-9]+)\s+((?:[a-z]|\:|\>|\!))\s+(.*)')
 
     def __init__(self, actor):