+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 =========
 
 * :support:`-` Rename ``clu_exchange`` to ``sdss_exchange``.
+* :support:`38` Use reply code ``>`` when the command is set to `~.CommandStatus.RUNNING`.
 
 * :release:`0.5.0 <2020-09-01>`
 * :support:`-` First version with >80% test coverage.